refactor: move gym_to_gymnasium and wrappers to utils directory
