+++ src/programmatic_policy_learning/envs/__init__.py
@@ -1,5 +1,5 @@
 """Public API for env utilities (adapters, helpers)"""
 
-from .gym_to_gymnasium import GymToGymnasium
+from .utils.gym_to_gymnasium import GymToGymnasium
 
 __all__ = ["GymToGymnasium"]
